refactor(pos-security-common): promote requiredPermissionsOperationCustomizer, drop 20 copies (#781)
The springdoc OperationCustomizer that publishes x-required-permissions was
copy-pasted byte-identically into all 20 services' OpenApiConfig. Promote it to
a single source in pos-security-common as an auto-configuration and delete the
20 local copies.

- RequiredPermissionsOpenApiAutoConfiguration: @AutoConfiguration gated by
  @ConditionalOnClass(OperationCustomizer.class) so non-web consumers are
  unaffected, with @ConditionalOnMissingBean so any lingering local customizer
  keeps precedence during migration. Registered via
  META-INF/spring/...AutoConfiguration.imports.
- pos-security-common gains an optional springdoc dependency (compile-only for
  this class; every consuming service already has springdoc at runtime).
- Each service's OpenApiConfig keeps its service-specific customOpenAPI() bean
  and drops the customizer + its private helpers + now-unused imports.

Contract-neutral by construction (identical logic). Verified: full reactor
compiles; the auto-configured customizer still emits x-required-permissions +
the AUTHENTICATED sentinel (pos-security-common test + a pos-order spec re-gen).
Completes the customizer-promotion half of #781; the mcp-server fail-closed
parse half shipped earlier.
